requirements.txt
pandas==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1
scipy==1.11.4
textblob==0.17.1
matplotlib==3.8.0
seaborn==0.13.0
numpy==1.26.4
